requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
contourpy==1.3.1
cpca==0.5.5
cycler==0.12.1
exceptiongroup==1.2.2
fonttools==4.55.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
loguru==0.7.3
matplotlib==3.10.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.0.0
propcache==0.2.1
pyahocorasick==2.1.0
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.8.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
-e git+ssh://git@github.com/leopeng1995/qwergpt.git@ad81cbbef13bacc3143c601048720f5f2adfbb54#egg=qwergpt
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==14.1
yarl==1.18.3
zhipuai==2.1.5.20241204
